[#388] Changed timers from milliseconds to nanoseconds

+++ sr_port/ionl_open.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2009 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2009 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -24,7 +27,7 @@
 
 LITREF unsigned char io_params_size[];
 
-short ionl_open(io_log_name *dev_name, mval *pp, int fd, mval *mspace, int4 timeout)
+short ionl_open(io_log_name *dev_name, mval *pp, int fd, mval *mspace, uint8 timeout)
 {
 	unsigned char	ch;
 	io_desc		*d_in, *d_out, *ioptr;
