+++ sr_port/ionl_rdone.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -14,7 +17,7 @@
 #include "mdef.h"
 #include "io.h"
 
-int	ionl_rdone(mint *val, int4 timeout)
+int	ionl_rdone(mint *val, uint8 timeout)
 {
 	mval	tmp;
