+++ sr_port/ionl_read.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -12,7 +15,7 @@
 #include "mdef.h"
 #include "io.h"
 
-int	ionl_read(mval *val, int4 timeout)
+int	ionl_read(mval *val, uint8 timeout)
 {
 	return ionl_readfl(val, 1, timeout);
 }
