+++ sr_port/ionl_readfl.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -14,7 +17,7 @@
 
 GBLREF io_pair io_curr_device;
 
-int ionl_readfl(mval *val, int4 width, int4 timeout)
+int ionl_readfl(mval *val, int4 width, uint8 timeout)
 {
 	io_desc *dv;
 	error_def(ERR_IOEOF);
