+++ sr_port/ious_dummy.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ * Copyright 2001 Sanchez Computer Associates, Inc.		*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -12,7 +15,7 @@
 #include "mdef.h"
 #include "io.h"
 
-short ious_dummy(io_log_name *dev_name, mval *pp, int fd, mval *mspace, int4 timeout)
+short ious_dummy(io_log_name *dev_name, mval *pp, int fd, mval *mspace, uint8 timeout)
 {
 	return 0;
 }
