+++ sr_unix/ious_rdone.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -15,7 +18,7 @@
 
 GBLREF io_pair		io_curr_device;
 
-int ious_rdone(mint *v, int4 t)
+int ious_rdone(mint *v, uint8 t)
 {
 	*v = -1;
 	return TRUE;
