+++ sr_unix/ious_readfl.c
@@ -1,6 +1,9 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ * Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *								*
+ * Copyright (c) 2019 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -17,7 +20,7 @@
 GBLREF spdesc		stringpool;
 GBLREF io_pair		io_curr_device;
 
-int ious_readfl(mval *v, int4 length, int4 t)
+int ious_readfl(mval *v, int4 length, uint8 t)
 {
 	v->str.len = 0;
 	v->str.addr = (char *)0;
